Change initial test function to a functor
This allows us to embed additional information about state (here the
bool mResult parameter) using a functor (functional operator) in an
object. Using operator overloading, we can use the object just like a
function by calling the instance (created as "test") as a function
(where we call "test()" in main()).

The parameter mResult is not being used yet, but will in future will be
used to provide the results of a test - pass or fail.