Extend to account for multiple tests
An interface is created to provide a means of collecting a number of
test cases. Whenever the "TEST" macro is called, a pointer to each
"Test" object is pushed into a vector. Now in main.cc the collection is
iterated over to execute each test case.

Currently, we only have one test case, however, we have things in place
to account for multiple test cases.