Catch exceptions in test failure
This requires a new method which means we need to make the base class
for tests concrete instead of pure virtual. To do this, we also move the
properties to this base class (name, passed flag) and introduce a new
one to provide information on why the test failed. These are made
private with methods used to allow public read access.

This simplifies the TEST macros slightly, and requires a change in the
test when accessing the name of the test (which is now private so not
accessible in the TES macro).