Allow for multiple tests in a test suite
To enable defining multiple tests in the test suite created in
Creation.cc, we need to ensure unique test instances. Before this change
each time the TEST macro was called, a Test class was declared and an
instance called "test" was created. Using macros, the creation of the
Test class and the test instance appends a line number - which is the
line number in the file where the TEST macro is called. This means each
time TEST is used it creates a unique Test class and test instance with
the line number appended - e.g., if TEST is called at line 129 of a
file, the test class will be Test129 and the test instance test129.