Account for missing exception when one expected
If the test case is looking for an exception but one is not thrown, then
capture this and report that outcome as failure with a relevant message.

This required us to define an exception for this case and then throw
this type of exception (MissingException) when no exception is thrown in
the test case.