Incorporate new test file which for confirm tests
This requires defining a namespace around the test class to prevent name
clashes with tests in different files. Without the namespace we have two
tests named "Test4" and instantiated as "test4" - since there is a TEST
macro on line 4 of both "Creation.cc" and "Confirm.cc". Wrapping the
classes in an unnamed namespace ensures there is no name clash when
linking.