Set expected failure message
This causes the tests to pass - since they are expected to fail (if
that even makes sense). We are testing that there is a failure and that
the failure message is as expected.

Also had to correct issues with TestBase class which defined the
expected failure message using incorrect type.