Add tests for floating point types
Need to override confirm() for each type to avoid ambiguity. Also, when
comparing floating point types need to allow small margin of error. This
is due to round-off associated with representing floating point types as
binary numbers.

This also exposed issue with long long integer tests which needed the
expected values to be explicitly typed by adding "LL" to the end of the
numbers.